pyproject.toml
[tool.poetry]
name = "task-management-backend"
version = "0.1.0"
description = "Task management API backend (FastAPI, SQLAlchemy, Pydantic, Supabase)"
authors = []
readme = "README.md"
packages = [{ include = "app", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
sqlalchemy = "^2.0.0"
pydantic = "^2.0.0"
alembic = "^1.14.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
supabase = "^2.0.0"
psycopg2-binary = "^2.9.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["src/tests"]
markers = [
    "integration: marks tests as integration tests (hit API and DB).",
]
